requirements.txt
airtest
pocoui
pydirectinput
pyautogui